+++ CHANGELOG.md
@@ -12,7 +12,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Show related persons on the organization detail page (a person is related to
   an organization if s.he is in the team of a course of this organization),
-- On a person detail page, show courses to which s.he participated.
+- On a person detail page, show courses to which s.he participated and
+  blogposts s.he authored.
 
 ### Changed